requirements.txt
python-telegram-bot>=20.0
aiohttp>=3.9
python-dotenv>=1.0
